requirements.txt
flask==2.3.3
werkzeug==2.3.7
jinja2==3.1.2
itsdangerous==2.1.2
click==8.1.7
markupsafe==2.1.3
python-dotenv==1.0.0

# App1 dependencies
openpyxl==3.1.2
python-docx==0.8.11

# App2 dependencies
pdfplumber==0.10.2
pandas==2.0.3
numpy==1.24.3

# App3 dependencies
docx2txt==0.8
shutil36==0.1.0
zipfile36==0.1.3 